% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBB_PE_gt_subset}
\alias{BBB_PE_gt_subset}
\title{Example for "gt" format specification}
\format{A data frame with 1,000 rows and 14 variables:
\describe{
\item{chr}{The chromosome number}
\item{pos}{The position of the marker}
\item{allele1}{The name of the first marker allele}
\item{allele2}{The name of the second marker allele}
\item{id1}{The genotypes for the first individuals (id1)}
\item{id2}{The genotypes for the second individuals (id2)}
\item{id3, id4, id5, id6, id7, id8, id9, id10}{The genotypes of
the remaining individuals}
}}
\usage{
BBB_PE_gt_subset
}
\description{
A dataset containing real genotypes for 1,000 SNPs.
The data is available for ten individuals and the first 1,000 SNPs on chromosome 1.
The data corresponds to a WGS experiment.
There is one columns per individual.
}
\keyword{datasets}
