% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBB_samples}
\alias{BBB_samples}
\title{A file with names or IDs for ten samples.}
\format{A data frame with 10 rows and one variable.
\describe{
\item{IDs}{The ids for ten individuals}
}}
\usage{
BBB_samples
}
\description{
The names (or IDs) are provided for ten samples.
}
\keyword{datasets}
