% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genoex}
\alias{genoex}
\title{Subset of a dataset with genotypes for 20 sheeps}
\format{
A data frame with 14,831 rows and 25 variables:
\describe{
  \item{chr}{The chromosome number}
  \item{marker_name}{The name of the marker}
  \item{pos}{The position of the marker}
  \item{allele1}{The name of the first marker allele}
  \item{allele2}{The name of the second marker allele}
  \item{id1}{The genotype for the first individual}
  \item{id2}{The genotype for the second individual}
  \item{id3, id4, id5, id6, id7, id8, id9, id10, id11, id12, id13, id14, id15, id16, id17, id18, id19, id20}{The genotypes of the remaining individuals} }
}
\usage{
genoex
}
\description{
A dataset containing real genotypes for 20 individuals. Genotypes are
available for 14,831 SNPs from the first three chromosomes were selected. The
twenty last columns correspond to the genotypes. Missing genotypes are set to
9.
}
\keyword{datasets}
