% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genosim}
\alias{genosim}
\title{Example from a small simulated data set}
\format{A data frame with 10,000 rows and 24 variables:
\describe{
  \item{chr}{The chromosome number}
  \item{pos}{The position of the marker}
  \item{allele1}{The name of the first marker allele}
  \item{allele2}{The name of the second marker allele}
  \item{id1}{The genotype for the first individual}
  \item{id2}{The genotype for the second individual}
  \item{id3, id4, id5, id6, id7, id8, id9, id10, id11, id12, id13, id14, id15, id16, id17, id18, id19, id20}{The genotypes of the remaining individuals}
 }}
\usage{
genosim
}
\description{
A dataset containing simulated genotypes for 20 individuals. Genotypes are
available for 10,000 SNPs on 10 chromosomes. The ten last columns correspond
to the genotypes.
}
\keyword{datasets}
