% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbd1}
\alias{hbd1}
\title{The result of local IBD probabilities for a pair of individuals among the 18
cows from the Amsterdam Island.}
\format{
the result is an array.
}
\usage{
hbd1
}
\description{
The results were obtained by running the a model with 4 layers and the predhbd
function for all HBD classes. Results were extracted for chromosome 1.
}
\details{
Results were extracted for chromosome 1 and the map is also provided (map1).
}
\keyword{datasets}
