% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbd2}
\alias{hbd2}
\title{The result of local IBD probabilities for a pair of individuals among the 18
cows from the Amsterdam Island (recent only).}
\format{
the result is an array.
}
\usage{
hbd2
}
\description{
The results were obtained by running the a model with 4 layers and the predhbd
function for HBD classes with a rate < 10. Results were extracted for
chromosome 1.
}
\keyword{datasets}
