% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kintaf_mix4l}
\alias{kintaf_mix4l}
\title{The result of a kinship analysis on 18 cattle from the Amsterdam Island.}
\format{
the results are a zres object.
}
\usage{
kintaf_mix4l
}
\description{
The results were obtained by running the a model with 4 layers and the kinship
option among all pairs of individuals (genotyped at 23679 SNPs after filtering).
}
\keyword{datasets}
