% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map1}
\alias{map1}
\title{The map for local IBD probabilities for a pair of individuals among the 18
cows from the Amsterdam Island (recent only).}
\format{
the result is an array.
}
\usage{
map1
}
\description{
The map corresponds to the results obtained by running the a model with 4
layers and the predhbd function (see hbd1 and hbd2). Results were extracted for
chromosome 1.
}
\keyword{datasets}
