% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooroh_accessor.R
\name{merge_zres}
\alias{merge_zres}
\title{Merge several zres objects generated by zoorun}
\usage{
merge_zres(list_zres)
}
\arguments{
\item{list_zres}{A list with the name of the zres objects to be merged.}
}
\value{
a single zres object containing the results from the merged zres
 objects. Note that the HBD segments are not sorted by ID number.
}
\description{
The function is used for example when the analysis has been split in several
tasks. All the zres must come from the same original
zoodata (zoorun is applied to the same zoodata with the same model). In
addition, the data should not overlap (the same individual should not be
present in multiple zres).
}
