% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rara_mix10r}
\alias{rara_mix10r}
\title{The result of an analysis on 22 sheeps from Rasa Aragonesa population.}
\format{
the results are a zres object.
}
\usage{
rara_mix10r
}
\description{
The results were obtained by running the default model (10 classes
 with pre-defined rates) on 22 individuals genotyped at 37465 SNPs.
}
\keyword{datasets}
