% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soay_mix10l}
\alias{soay_mix10l}
\title{The result of an analysis on 110 sheeps from the Soay population.}
\format{
the results are a zres object.
}
\usage{
soay_mix10l
}
\description{
The results were obtained by running the default model (10 layers
 with pre-defined rates) on 110 individuals genotyped at 37465 SNPs.
}
\keyword{datasets}
