% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soay_mix10r}
\alias{soay_mix10r}
\title{The result of an analysis on 110 sheeps from the Soay population.}
\format{the results are a zres object.}
\usage{
soay_mix10r
}
\description{
The results were obtained by running the default model (10 classes
 with pre-defined rates) on 110 individuals genotyped at 37465 SNPs.
}
\keyword{datasets}
