% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typs}
\alias{typs}
\title{Subset of a dataset with genotypes for 6 individuals from a cattle population.}
\format{A data frame with 6370 rows and 10 variables:
\describe{
\item{chr}{The chromosome number}
\item{pos}{The position of the marker}
\item{allele1}{The name of the first marker allele}
\item{allele2}{The name of the second marker allele}
\item{id1}{The genotypes for the first individuals (id1)}
\item{id2}{The genotypes for the second individuals (id2)}
\item{id3}{The genotypes for the third individuals (id3)}
\item{id4}{The genotypes for the fourth individuals (id4)}
\item{id5}{The genotypes for the fifth individuals (id5)}
\item{id6}{The genotypes for the sixth individuals (id6)}
}}
\usage{
typs
}
\description{
A dataset containing real genotypes for 6 individuals.
Genotypes are available for a low density array with 6370 SNPs on 29 autosomes.
The six last columns correspond to the genotypes.
Missing genotypes are set to 9.
}
\keyword{datasets}
