% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typsfrq}
\alias{typsfrq}
\title{A file with marker allele frequencies for the cattle population.}
\format{
A data frame with 6370 rows and one variable.
\describe{
\item{frq}{The allele frequencies estimated for the first allele}
}
}
\usage{
typsfrq
}
\description{
The allele frequencies of the first allele were computed in a larger sample.
}
\keyword{datasets}
