% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wilt_mix10l}
\alias{wilt_mix10l}
\title{The result of an analysis on 23 sheeps from Wiltshire population.}
\format{
the results are a zres object.
}
\usage{
wilt_mix10l
}
\description{
The results were obtained by running the default model (10 layers
 with pre-defined rates) on 23 individuals genotyped at 37465 SNPs.
}
\keyword{datasets}
