% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wilt_mix10r}
\alias{wilt_mix10r}
\title{The result of an analysis on 23 sheeps from Wiltshire population.}
\format{the results are a zres object.}
\usage{
wilt_mix10r
}
\description{
The results were obtained by running the default model (10 classes
 with pre-defined rates) on 23 individuals genotyped at 37465 SNPs.
}
\keyword{datasets}
